@@ -1,7 +1,7 @@
 {
   "name": "morgan",
   "description": "HTTP request logger middleware for node.js",
-  "version": "1.9.1",
+  "version": "1.10.0",
   "contributors": [
     "Douglas Christopher Wilson <doug@somethingdoug.com>",
     "Jonathan Ong <me@jongleberry.com> (http://jongleberry.com)"
@@ -15,24 +15,24 @@
   ],
   "repository": "expressjs/morgan",
   "dependencies": {
-    "basic-auth": "~2.0.0",
+    "basic-auth": "~2.0.1",
     "debug": "2.6.9",
-    "depd": "~1.1.2",
+    "depd": "~2.0.0",
     "on-finished": "~2.3.0",
-    "on-headers": "~1.0.1"
+    "on-headers": "~1.0.2"
   },
   "devDependencies": {
-    "eslint": "5.5.0",
-    "eslint-config-standard": "12.0.0",
-    "eslint-plugin-import": "2.14.0",
-    "eslint-plugin-markdown": "1.0.0-beta.6",
-    "eslint-plugin-node": "7.0.1",
-    "eslint-plugin-promise": "4.0.1",
-    "eslint-plugin-standard": "4.0.0",
-    "istanbul": "0.4.5",
-    "mocha": "2.5.3",
+    "eslint": "6.8.0",
+    "eslint-config-standard": "14.1.0",
+    "eslint-plugin-import": "2.20.1",
+    "eslint-plugin-markdown": "1.0.2",
+    "eslint-plugin-node": "9.2.0",
+    "eslint-plugin-promise": "4.2.1",
+    "eslint-plugin-standard": "4.0.1",
+    "mocha": "7.1.1",
+    "nyc": "15.0.0",
     "split": "1.0.1",
-    "supertest": "1.1.0"
+    "supertest": "4.0.2"
   },
   "files": [
     "LICENSE",
@@ -46,7 +46,7 @@
   "scripts": {
     "lint": "eslint --plugin markdown --ext js,md .",
     "test": "mocha --check-leaks --reporter spec --bail",
-    "test-cov": "istanbul cover node_modules/mocha/bin/_mocha -- --check-leaks --reporter dot",
-    "test-travis": "istanbul cover node_modules/mocha/bin/_mocha --report lcovonly -- --check-leaks --reporter spec"
+    "test-ci": "nyc --reporter=text npm test",
+    "test-cov": "nyc --reporter=html --reporter=text npm test"
   }
 }
